updated to support clients with multiple PIDs
Probably should not have done it this way.  The original bug was with 'exit' command, only one client PID was getting the command and the server would exit.  Now, the server will wait until all the seen PIDs also get the command.  But really the solution should be that if ANY PID on the client sees the exit command, then abort all processes.  Cleaner.